Update Colors and Game Desc
The button and link colors along with the footer, homepage, and all text colors were updated. Text is no longer black on white, instead its cyan-900 on white. Homepage background is now white again. The footer is the blue-900 that was the homepage background previously because it looks cleaner. Media/Press Kit page was updated with added facts and links along with the download being renamed to kit.zip.